+++ contrib/platform/src/com/sun/jna/platform/win32/Kernel32.java
@@ -1414,6 +1414,71 @@ public interface Kernel32 extends StdCallLibrary, WinNT, Wincon {
     boolean GetVersionEx(OSVERSIONINFOEX lpVersionInfo);
 
     /**
+     * The VerifyVersionInfoW function compares a set of operating system
+     * version requirements to the corresponding values for the currently
+     * running version of the system.This function is subject to manifest-based
+     * behavior.
+     * 
+     * @param lpVersionInformation
+     *            A pointer to an OSVERSIONINFOEX structure containing the
+     *            operating system version requirements to compare. The
+     *            dwTypeMask parameter indicates the members of this structure
+     *            that contain information to compare.
+     * 
+     *            You must set the dwOSVersionInfoSize member of this structure
+     *            to sizeof(OSVERSIONINFOEX). You must also specify valid data
+     *            for the members indicated by dwTypeMask. The function ignores
+     *            structure members for which the corresponding dwTypeMask bit
+     *            is not set.
+     * @param dwTypeMask
+     *            A mask that indicates the members of the OSVERSIONINFOEX
+     *            structure to be tested.
+     * @param dwlConditionMask
+     *            The type of comparison to be used for each lpVersionInfo
+     *            member being compared. To build this value, call the
+     *            VerSetConditionMask function once for each OSVERSIONINFOEX
+     *            member being compared.
+     * @return If the currently running operating system satisfies the specified
+     *         requirements, the return value is a nonzero value.
+     * 
+     *         If the current system does not satisfy the requirements, the
+     *         return value is zero and GetLastError returns
+     *         ERROR_OLD_WIN_VERSION.
+     * 
+     *         If the function fails, the return value is zero and GetLastError
+     *         returns an error code other than ERROR_OLD_WIN_VERSION.
+     */
+    boolean VerifyVersionInfoW(OSVERSIONINFOEX lpVersionInformation, int dwTypeMask, long dwlConditionMask);
+
+    /**
+     * The VerSetConditionMask function sets the bits of a 64-bit value to
+     * indicate the comparison operator to use for a specified operating system
+     * version attribute. This function is used to build the dwlConditionMask
+     * parameter of the VerifyVersionInfo function.
+     * 
+     * @param conditionMask
+     *            A value to be passed as the dwlConditionMask parameter of the
+     *            VerifyVersionInfo function. The function stores the comparison
+     *            information in the bits of this variable.
+     * 
+     *            Before the first call to VerSetCondition, initialize this
+     *            variable to zero. For subsequent calls, pass in the variable
+     *            used in the previous call.
+     * @param typeMask
+     *            A mask that indicates the member of the OSVERSIONINFOEX
+     *            structure whose comparison operator is being set. This value
+     *            corresponds to one of the bits specified in the dwTypeMask
+     *            parameter for the VerifyVersionInfo function.
+     * @param condition
+     *            The operator to be used for the comparison. The
+     *            VerifyVersionInfo function uses this operator to compare a
+     *            specified attribute value to the corresponding value for the
+     *            currently running system.
+     * @return The function returns the condition mask value.
+     */
+    long VerSetConditionMask(long conditionMask, int typeMask, byte condition);
+
+    /**
      * The GetSystemInfo function returns information about the current system.
      *
      * @param lpSystemInfo