+++ contrib/platform/src/com/sun/jna/platform/win32/WinNT.java
@@ -1988,13 +1988,13 @@ public interface WinNT extends WinError, WinDef, WinBase, BaseTSD {
         }
     }
 
-    int VER_EQUAL = 1;
-    int VER_GREATER = 2;
-    int VER_GREATER_EQUAL = 3;
-    int VER_LESS = 4;
-    int VER_LESS_EQUAL = 5;
-    int VER_AND = 6;
-    int VER_OR = 7;
+    byte VER_EQUAL = 1;
+    byte VER_GREATER = 2;
+    byte VER_GREATER_EQUAL = 3;
+    byte VER_LESS = 4;
+    byte VER_LESS_EQUAL = 5;
+    byte VER_AND = 6;
+    byte VER_OR = 7;
 
     int VER_CONDITION_MASK = 7;
     int VER_NUM_BITS_PER_CONDITION_MASK = 3;
@@ -2015,6 +2015,23 @@ public interface WinNT extends WinError, WinDef, WinBase, BaseTSD {
     int VER_PLATFORM_WIN32s = 0;
     int VER_PLATFORM_WIN32_WINDOWS = 1;
     int VER_PLATFORM_WIN32_NT = 2;
+
+    /*
+     * WIN32_WINNT version constants
+     */
+    short WIN32_WINNT_NT4 = 0x0400; // Windows NT 4.0
+    short WIN32_WINNT_WIN2K = 0x0500; // Windows 2000
+    short WIN32_WINNT_WINXP = 0x0501; // Windows XP
+    short WIN32_WINNT_WS03 = 0x0502; // Windows Server 2003
+    short WIN32_WINNT_WIN6 = 0x0600; // Windows Vista
+    short WIN32_WINNT_VISTA = 0x0600; // Windows Vista
+    short WIN32_WINNT_WS08 = 0x0600; // Windows Server 2008
+    short WIN32_WINNT_LONGHORN = 0x0600; // Windows Vista
+    short WIN32_WINNT_WIN7 = 0x0601; // Windows 7
+    short WIN32_WINNT_WIN8 = 0x0602; // Windows 8
+    short WIN32_WINNT_WINBLUE = 0x0603; // Windows 8.1
+    short WIN32_WINNT_WINTHRESHOLD = 0x0A00; // Windows 10
+    short WIN32_WINNT_WIN10 = 0x0A00; // Windows 10
 
     /**
      * Read the records sequentially. If this is the first read operation, the